#278 [refactor] 뒤로가기 버튼 클릭하는 로직 확장함수 구현
